refactor: move TPF resource_tests to sdkv2
