+++ internal/service/advancedclustertpf/move_state_test.go
@@ -9,6 +9,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/tfversion"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/testutil/acc"
 )
+const (
+	resourceName         = "mongodbatlas_advanced_cluster.test"
+)
 
 func TestAccAdvancedCluster_moveBasic(t *testing.T) {
 	var (
